+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.0.1-dev.3
+- ✨ Add stack trace to `onError`
+- 🐛 Fix debug delay not working on chained jobs
+
 ## 0.0.1-dev.2
 - 🐛 Fix `deriveOnPresent` typo